+++ R/comparedf.control.R
@@ -3,7 +3,7 @@
 #'
 #' Control tolerance definitions for the \code{\link{comparedf}} function.
 #'
-#' @param tol.logical,tol.num,tol.char,tol.factor,tol.date,tol.other A function or one of the shortcut character strings,
+#' @param tol.logical,tol.num,tol.char,tol.factor,tol.date,tol.other A function or one of the shortcut character strings or a list thereof,
 #'   denoting the tolerance function to use for a given data type. See "details", below.
 #' @param tol.num.val Numeric; maximum value of differences allowed in numerics (fed to the function given in \code{tol.num}).
 #' @param int.as.num Logical; should integers be coerced to numeric before comparison? Default FALSE.
@@ -39,6 +39,9 @@
 #'   \item{\code{tol.other = "none"}: expect objects of other classes to be exactly identical.}
 #' }
 #'
+#' A list with names mapped to \code{x} can be used to specify tolerances by variable. One unnamed element is supported
+#' as the default.
+#'
 #' \code{tol.vars}: If not set to \code{"none"} (the default) or a named vector,
 #'   the \code{tol.vars} argument is a character vector denoting equivalence classes
 #'   for the characters in the variable names. A single character in this vector means to replace that character
@@ -55,6 +58,12 @@
 #'                "._",   # set all underscores to dots.
 #'                "e")    # remove all letter e's
 #' )
+#'
+#' comparedf.control(tol.char = list(
+#'   "none",      # the default
+#'   x1 = "case", # be case-insensitive for the variable "x1"
+#'   x2 = function(x, y) tol.NA(x, y, x != y | y == "NA") # a custom-defined tolerance
+#' ))
 #' @seealso \code{\link{comparedf}}, \code{\link{comparedf.tolerances}}, \code{\link{summary.comparedf}}
 #' @author Ethan Heinzen
 #' @export
@@ -73,26 +82,40 @@ comparedf.control <- function(
   max.print.vars = NA, max.print.obs = NA, max.print.diffs.per.var = 10,
   max.print.diffs = 50, max.print.attrs = NA, ..., max.print.diff = 10)
 {
+  make_tol <- function(x, prefix, defaults)
+  {
+    if(is.function(x)) return(list(x))
+    if(is.character(x)) return(list(match.fun(paste0(prefix, match.arg(x, defaults, several.ok = FALSE)))))
+    if(!is.list(x) || (length(x) > 1 && (is.null(names(x)) || sum(names(x) == "") > 1)))
+      stop("Provided tolerances should be a function, single character string, or a named list of functions/character strings")
+
+    out <- lapply(x, function(y) {
+      if(!is.function(y)) match.fun(paste0(prefix, match.arg(y, defaults, several.ok = FALSE))) else y
+    })
+    names(out) <- names(x)
+    out
+  }
+
   #### Logical ####
-  if(!is.function(tol.logical)) tol.logical <- match.fun(paste0("tol.logical.", match.arg(tol.logical, several.ok = FALSE)))
+  tol.logical <- make_tol(tol.logical, "tol.logical.", "none")
 
   #### Numerics ####
   if(!is.numeric(tol.num.val)) stop("'tol.num.val' needs to be numeric.")
-  if(!is.function(tol.num)) tol.num <- match.fun(paste0("tol.num.", match.arg(tol.num, several.ok = FALSE)))
+  tol.num <- make_tol(tol.num, "tol.num.", c("absolute", "percent", "pct"))
   if(!is.logical(int.as.num) || length(int.as.num) != 1 || is.na(int.as.num)) stop("'int.as.num' should be TRUE or FALSE.")
 
   #### Characters and factors ####
-  if(!is.function(tol.char)) tol.char <- match.fun(paste0("tol.char.", match.arg(tol.char, several.ok = FALSE)))
+  tol.char <- make_tol(tol.char, "tol.char.", c("none", "trim", "case", "both"))
 
-  if(!is.function(tol.factor)) tol.factor <- match.fun(paste0("tol.factor.", match.arg(tol.factor, several.ok = FALSE)))
+  tol.factor <- make_tol(tol.factor, "tol.factor.", c("none", "levels", "labels"))
   if(!is.logical(factor.as.char) || length(factor.as.char) != 1 || is.na(factor.as.char)) stop("'factor.as.char' should be TRUE or FALSE.")
 
   #### Dates ####
   if(!is.numeric(tol.date.val)) stop("'tol.date.val' needs to be numeric.")
-  if(!is.function(tol.date)) tol.date <- match.fun(paste0("tol.date.", match.arg(tol.date, several.ok = FALSE)))
+  tol.date <- make_tol(tol.date, "tol.date.", "absolute")
 
   #### Other ####
-  if(!is.function(tol.other)) tol.other <- match.fun(paste0("tol.other.", match.arg(tol.other, several.ok = FALSE)))
+  tol.other <- make_tol(tol.other, "tol.other.", "none")
 
   #### Variable names ####
   if(!is.character(tol.vars)) stop("'tol.vars' must be a character string or vector.")